+++ example/main.dart
@@ -1,22 +1,23 @@
-import 'package:validifyDart/validifyDart.dart';
+
+import 'package:validifydart/validify_dart.dart';
 
 Future<void> main() async {
-  print(validifyDart.isValidEmail("rudradeep@gmail.com"));  // true
-  print(validifyDart.isConnectedToInternet());  // true/falseimport 'package:validifyDart/validifyDart.dart';
+  print(ValidifyDart.isValidEmail("rudradeep@gmail.com")); // true
+  print(ValidifyDart
+      .isConnectedToInternet()); // true
 
 // Validating phone number
-  bool isPhoneValid = validifyDart.isValidPhone('+1234567890');
+  bool isPhoneValid = ValidifyDart.isValidPhone('+1234567890');
 
 // Validating Aadhaar number
-  bool isAadhaarValid = validifyDart.isValidAadhaar('123456789012');
+  bool isAadhaarValid = ValidifyDart.isValidAadhaar('123456789012');
 
 // Checking if connected to the internet
-  bool isConnected = await validifyDart.isConnectedToInternet();
+  bool isConnected = await ValidifyDart.isConnectedToInternet();
 
 // Checking if a number is prime
-  bool isPrime = validifyDart.isPrime(7);
+  bool isPrime = ValidifyDart.isPrime(7);
 
 // Validating email
-  bool isEmailValid = validifyDart.isValidEmail('rudradeep@gmail.com');
-
+  bool isEmailValid = ValidifyDart.isValidEmail('rudradeep@gmail.com');
 }